adding -beta or -experimental to services when appropriate. Fixes #43
