+++ examples/relationship_extraction.v1-beta.js
@@ -5,7 +5,7 @@ var watson = require('watson-developer-cloud');
 var relationship_extraction = watson.relationship_extraction({
   username: 'INSERT YOUR USERNAME FOR THE SERVICE HERE',
   password: 'INSERT YOUR PASSWORD FOR THE SERVICE HERE',
-  version: 'v1'
+  version: 'v1-beta'
 });
 
 relationship_extraction.extract({
