+++ examples/visual_insights.v1-experimental.js
@@ -6,7 +6,7 @@ var fs = require('fs');
 var visual_insights = watson.visual_insights({
   username: 'INSERT YOUR USERNAME FOR THE SERVICE HERE',
   password: 'INSERT YOUR PASSWORD FOR THE SERVICE HERE',
-  version: 'v1'
+  version: 'v1-experimental'
 });
 
 var params = {
