+++ examples/visual_recognition.v1-beta.js
@@ -6,7 +6,7 @@ var fs = require('fs');
 var visual_recognition = watson.visual_recognition({
   username: 'INSERT YOUR USERNAME FOR THE SERVICE HERE',
   password: 'INSERT YOUR PASSWORD FOR THE SERVICE HERE',
-  version: 'v1'
+  version: 'v1-beta'
 });
 
 var params = {
