+++ lib/index.js
@@ -17,6 +17,7 @@
 'use strict';
 
 var path         = require('path');
+var fs           = require('fs');
 var util         = require('util');
 var extend       = require('extend');
 var vcapServices = require('vcap_services');
@@ -93,7 +94,16 @@ function createServiceAPI(serviceName) {
       return Object.freeze(s);
     } catch (e) {
       if (e.code === 'MODULE_NOT_FOUND') {
-        throw new Error(util.format('Service %s %s not found', serviceName, version));
+        // for the v1.0 breaking change of requiring experimental/beta in the module name, try to offer a helpful error message.
+        // https://github.com/watson-developer-cloud/node-sdk/issues/43
+        var msg = 'Service %s %s not found.';
+        if (fs.existsSync(path.join(__dirname, '..', 'services', serviceName, path.basename(version) + '-beta.js'))) {
+          msg += ' Did you mean ' + path.basename(version) + '-beta?';
+        }
+        if (fs.existsSync(path.join(__dirname, '..', 'services', serviceName, path.basename(version) + '-experimental.js'))) {
+          msg += ' Did you mean ' + path.basename(version) + '-experimental?';
+        }
+        throw new Error(util.format(msg, serviceName, version));
       } else {
         throw e;
       }
