+++ services/visual_recognition/v1-beta.js
