+++ test/test.concept_expansion.v1-beta.js
@@ -33,7 +33,7 @@ describe('concept_expansion', function() {
     username: 'batman',
     password: 'bruce-wayne',
     url: 'http://ibm.com:80',
-    version: 'v1'
+    version: 'v1-beta'
   };
 
   before(function() {
