+++ test/test.question_and_answer.v1-beta.js
@@ -10,7 +10,7 @@ describe('question_and_answer', function() {
     username: 'batman',
     password: 'bruce-wayne',
     url: 'http://ibm.com:80',
-    version: 'v1'
+    version: 'v1-beta'
   };
 
   var question = 'what is watson?';
