+++ test/test.tone_analyzer.v2-experimental.js
@@ -3,8 +3,9 @@
 var assert = require('assert');
 var watson = require('../lib/index');
 var nock   = require('nock');
+var qs = require('qs');
 
-describe('tone_analyzer.v1', function() {
+describe('tone_analyzer.v2', function() {
 
   var noop = function() {};
 
@@ -12,22 +13,21 @@ describe('tone_analyzer.v1', function() {
     text: 'IBM Watson Developer Cloud'
   };
   var synonym_request = {
-	  'words': ['greece'],
-	  'limit': 4
+	  'word': 'product'
   };
   var tone_response = {
     tree: {}
   };
   var synonym_response = [ { headword: 'x' } ];
 
-  var tone_path = '/v1/tone',
-	synonym_path = '/v1/synonym';
+  var tone_path = '/v2/tone',
+	synonym_path = '/v2/synonyms';
 
   var service = {
     username: 'batman',
     password: 'bruce-wayne',
     url: 'http://ibm.com:86',
-    version: 'v1'
+    version: 'v2-experimental'
   };
 
   before(function() {
@@ -36,7 +36,7 @@ describe('tone_analyzer.v1', function() {
       .persist()
       .post(tone_path, tone_request.text)
       .reply(200, tone_response)
-      .post(synonym_path, synonym_request)
+      .get(synonym_path + '?' + qs.stringify(synonym_request))
       .reply(200, synonym_response);
   });
 
@@ -87,11 +87,10 @@ describe('tone_analyzer.v1', function() {
 
   it('synonym API should generate a valid payload with content', function() {
       var req = tone_analyzer.synonym(synonym_request, noop);
-      var body = new Buffer(req.body).toString('ascii');
-      assert.equal(req.uri.href, service.url + synonym_path);
-      assert.equal(body, JSON.stringify(synonym_request));
-      assert.equal(req.method, 'POST');
-      assert.equal(req.headers['content-type'], 'application/json');
+      var myquery = qs.stringify(synonym_request);
+      assert.equal(req.uri.href, service.url + synonym_path + '?' + myquery);
+      assert.equal(req.uri.query, myquery);
+      assert.equal(req.method, 'GET');
   });
   it('synonym API should format the response', function(done) {
     tone_analyzer.synonym(synonym_request, function(err, response) {
